requirements.txt
Django==4.2
factory-boy==3.3.0
Faker==19.3.0
Pillow==9.5.0
pytest==7.4.0
